Ajustes gerais.

+++ app/Http/Controllers/API/UserController.php
@@ -94,6 +94,8 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isSecretaria');
+
         $user = User::findOrFail( $id );
         $user->delete();
